+++ app.rb
@@ -4,89 +4,19 @@ require './teacher'
 require './class_room'
 require './book'
 require './rental'
+require './book_module'
+require './person_module'
 
 class App
   def initialize
     @people = []
-    @books = []
     @rentals = []
+    @books = []
     @classroom = Classroom.new('Grade 10')
   end
 
-  def list_all_books
-    if @books.empty?
-      puts 'Oops! the library is empty, please add books'
-      return
-    end
-    @books.each do |book|
-      puts "Title: #{book.title}, Author: #{book.author}"
-    end
-  end
-
-  def list_all_people
-    if @people.empty?
-      puts 'Oops! no people registered, please add people'
-    else
-      @people.each do |person|
-        puts "[#{person.class}] Name: #{person.name}, ID: #{person.id}, Age: #{person.age}"
-      end
-    end
-  end
-
-  def create_person
-    puts 'Do you want to create a student (1) or a teacher(2)?....'
-    option = gets.chomp
-
-    case option
-    when '1'
-      create_student
-    when '2'
-      create_teacher
-    else
-      puts 'Invalid input! Please type 1 or 2'
-    end
-  end
-
-  def create_student
-    print 'Name:'
-    name = gets.chomp
-
-    print 'Age:'
-    age = gets.chomp.to_i
-
-    print 'Has Parent permission? [Y/N]:'
-    parent_permission = gets.chomp.downcase
-
-    student = Student.new(name, @classroom, age, parent_permission: parent_permission == 'y')
-    @people << student
-    puts 'Student created successfully'
-  end
-
-  def create_teacher
-    print 'Name:'
-    name = gets.chomp
-
-    print 'Age:'
-    age = gets.chomp.to_i
-
-    print 'Specialization:'
-    specialization = gets.chomp
-
-    @people << Teacher.new(name, specialization, age)
-    puts 'Teacher created successfully'
-  end
-
-  def create_book
-    print 'Title: '
-    title = gets.chomp
-
-    print 'Author: '
-    author = gets.chomp
-
-    @books << Book.new(title, author)
-
-    puts 'Book created successfully'
-  end
+  include BookModule
+  include PersonModule
 
   def create_rental
     puts 'Select a book from the following list by number'
