+++ book_module.rb
@@ -1,0 +1,23 @@
+module BookModule
+  def create_book
+    print 'Title: '
+    title = gets.chomp
+
+    print 'Author: '
+    author = gets.chomp
+
+    @books << Book.new(title, author)
+
+    puts 'Book created successfully'
+  end
+
+  def list_all_books
+    if @books.empty?
+      puts 'Oops! the library is empty, please add books'
+      return
+    end
+    @books.each do |book|
+      puts "Title: #{book.title}, Author: #{book.author}"
+    end
+  end
+end
